break out emissions line chart
